Added Links to Rails Girls
This is as a response to an email we got from someone who was confused about what an 'introductory workshop' is... Since they may be hearing about SoC having never heard of Rails Girls, then we should make sure they know where to look. I would also like to add a link to 'or similar' soon with an explanatory text on how they can get up to speed, in conjunction with our guide on what constitutes a beginner with @katrin-k and the other participants from last year.