fix navigation for mobile
I removed a div class 'container-fluid' which was making things very strange on mobile. I did this because the gutter width didn't have a unit value. When applied as a unit value of 30px, the 'container-fluid' was creating a 15px white space gutter on either side the page, so I removed that div as it was structurally unneccessary.