Update blog post, campaign and press pages
I have Included the link for the video (to anyone who sees this, please
don’t share the post nor the video until we start posting on twitter
and facebook).
The Campaign page also had the video, so it is updated. Besides this, I
made some changes to the Press Page so that we can publish it if you
feel like it’s ok.