Add sections and materials
Sections and materials:
* Press Releases 3/4
* Videos
* Quotes
* Press Coverage links

I also made a new heading “Press Releases and Media” to separate the
first part “About” with the rest. I pushed the link to download the
press kit to this section and wrote an intro. See if that’s ok or if we
need to change it.